Making some big changes. Got the api's refactored out to a new file.
