+++ web/index.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="studentData.css">
 </head>
 
-<body >
+<body ng-controller="DataTableController as StuCntrl" ng-cloak>
 <div class="container">
     
     <div class="jumbotron text-center">
@@ -66,7 +66,7 @@
 <!--======================================================-->
 
     <!-------- Table View for Student Data -------->
-    <div id="table-view" class="main-view table-responsive" ng-controller="DataTableController as StuCntrl" ng-cloak>
+    <div id="table-view" class="main-view table-responsive" >
         <table class="table table-bordered table-hover">
             <thead class="thead-color">
                 <tr>
@@ -364,37 +364,13 @@
 
 
     <!-------- Modal Student View for Displaying Student Data  -------->
-    <div class="modal fade" id="studentModal" role="dialog" ng-controller="StudentModalController as StuModalCtrl">
+    <div class="modal fade" id="studentModal" role="dialog">
         <div id="showStudentModal" class="modal-dialog">
             <!-- data is dynamically created in a script -->
-
-            <div class="modal-content text-center">
-                <div class="modal-header">
-                    <h4 class="modal-title"> {{selectedStudent.lname}}, {{selectedStudent.fname}}</h4>
-                </div>
-                <div class="modal-body">
-                    <div class="panel-body">
-                        <div id="studentModalDataFrame">
-                            <div class="studentModalData">
-                                <img id="modalPicture"
-                                     src="/student-images/{{selectedStudent.lname}}-{{selectedStudent.fname}}.png"
-                                     class="studentPicture">
-                                <p><b>Year: </b>{{selectedStudent.s_year}}</p>
-                                <p><b>Start Date: </b>{{selectedStudent.startDate}}</p>
-                                <p><b>Phone: </b>{{selectedStudent.phone}}</p>
-                                <p><b>Address: </b><br/>{{selectedStudent.street}}<br/>
-                                    {{selectedStudent.city}}, {{selectedStudent.state}} {{selectedStudent.zip}}</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                </div>
-            </div>
-
+            <student-info-modal></student-info-modal>
         </div>
     </div>
+
 
 
     <!------- Spinning Loading Modal  -------->
