+++ web/student_logic.js
@@ -8,10 +8,11 @@ myapp.service('classroomSrvc', ['$http', '$q',function ($http, $q) {
     let _students = [];
     let _manifest = [];
     let _getStudentCalls = [];
-    this.selectedStu = {};
-
+    let selectedStu = {};
+    //
     $http.get('/api/v1/students.json')
         .then(function(res) {
+            console.log(res.data);
             // For now just get the entire student list. We will learn to paginate with it.
             // TODO: learn to do server-side paging with MySQL
             for (stu in res.data) {
@@ -31,11 +32,12 @@ myapp.service('classroomSrvc', ['$http', '$q',function ($http, $q) {
     this.setSelectedStu = function(stu) {
         console.log("setting student");
         _selectedStu = stu;
+        console.log(_selectedStu);
     };
-    // this.getSelectedStu = function() {
-    //     console.log("getting student");
-    //     return _students[4];
-    // };
+    this.getSelectedStu = function() {
+        console.log("getting student");
+        return _students[4];
+    };
 
 
     this.getAll = function () {
@@ -51,7 +53,7 @@ myapp.controller('DataTableController', ['$scope', 'classroomSrvc', function ($s
 
 
     $scope.popStudentModal = function(stu) {
-        classroomSrvc.setSelectedStu(stu);
+        $scope.selectedStu = stu;
     };
 
     $scope.editButton = () => {
@@ -59,6 +61,13 @@ myapp.controller('DataTableController', ['$scope', 'classroomSrvc', function ($s
     };
 }]);
 
-myapp.controller('StudentModalController', ['$scope', 'classroomSrvc', function($scope, classroomSrvc) {
-    $scope.selectedStu = classroomSrvc.selectedStu;
-}]);
+// myapp.controller('StudentModalController', ['$scope', 'classroomSrvc', function($scope, classroomSrvc) {
+//     $scope.selectedStudent = classroomSrvc.getSelectedStu();
+// }]);
+
+myapp.directive('studentInfoModal', function() {
+    return {
+        restrict: 'E',
+        templateUrl: 'stu-modal.html'
+    };
+});
