--- web/students.json
@@ -1,314 +1,0 @@
-[
-	{
-		"fname": "Lois",
-		"lname": "Hanson",
-		"startDate": "3/12/93",
-		"street": "923 North 1400 East",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "673-3114",
-		"s_year": 4,
-		"id": "0001"
-	},
-	{
-		"fname": "Alex",
-		"lname": "Peterson",
-		"startDate": "7/11/94",
-		"street": "931 South 300 West",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "679-2116",
-		"s_year": 3,
-		"id": "0002"
-	},
-	{
-		"fname": "Elizabeth",
-		"lname": "Howard",
-		"startDate": "3/19/94",
-		"street": "1726 East 1330 North",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93679,
-		"phone": "734-3219",
-		"s_year": 3,
-		"id": "0003"
-	},
-	{
-		"fname": "Emmelie",
-		"lname": "Well",
-		"startDate": "12/23/92",
-		"street": "2710 West 3700 North",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93667,
-		"phone": "369-3334",
-		"s_year": 4,
-		"id": "0004"
-	},
-	{
-		"fname": "Anna",
-		"lname": "Ballicki",
-		"startDate": "9/14/93",
-		"street": "2710 West 3700 North",
-		"city": "Purdie",
-		"state": "ID",
-		"zip": 93667,
-		"phone": "369-3334",
-		"s_year": 3,
-		"id": "0005"
-	},
-	{
-		"fname": "Daniel",
-		"lname": "McCormic",
-		"startDate": "3/29/93",
-		"street": "44 West 300 South",
-		"city": "Greenville",
-		"state": "ID",
-		"zip": 93639,
-		"phone": "739-2131",
-		"s_year": 4,
-		"id": "0006"
-	},
-	{
-		"fname": "Joseph",
-		"lname": "Read",
-		"startDate": "4/19/94",
-		"street": "92 South 200 East",
-		"city": "Green Park",
-		"state": "ID",
-		"zip": 93619,
-		"phone": "369-3629",
-		"s_year": 3,
-		"id": "0007"
-	},
-	{
-		"fname": "Jake",
-		"lname": "Calvert",
-		"startDate": "10/27/92",
-		"street": "103 South 1230 East",
-		"city": "Blythville",
-		"state": "ID",
-		"zip": 93699,
-		"phone": "739-3171",
-		"s_year": 4,
-		"id": "0008"
-	},
-	{
-		"fname": "Jake",
-		"lname": "Paterson",
-		"startDate": "12/3/93",
-		"street": "103 South 1230 East",
-		"city": "Blythville",
-		"state": "ID",
-		"zip": 93699,
-		"phone": "739-3171",
-		"s_year": 3,
-		"id": "0009"
-	},
-	{
-		"fname": "Sara",
-		"lname": "Calvert",
-		"startDate": "7/12/94",
-		"street": "371 North 1420 East",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "734-1097",
-		"s_year": 3,
-		"id": "0010"
-	},
-	{
-		"fname": "Jared",
-		"lname": "Anderson",
-		"startDate": "11/23/92",
-		"street": "706 North 300 East",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "793-3129",
-		"s_year": 4,
-		"id": "0011"
-	},
-	{
-		"fname": "So-Ri",
-		"lname": "Choi",
-		"startDate": "6/4/94",
-		"street": "706 North 300 East",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "793-3129",
-		"s_year": 3,
-		"id": "0012"
-	},
-	{
-		"fname": "Edward",
-		"lname": "Hashimoto",
-		"startDate": "7/27/93",
-		"street": "1237 East Orchard Heights",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "733-0469",
-		"s_year": 4,
-		"id": "0013"
-	},
-	{
-		"fname": "Joe",
-		"lname": "Larson",
-		"startDate": "2/21/94",
-		"street": "1237 Island Dr. #2",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "679-1629",
-		"s_year": 3,
-		"id": "0014"
-	},
-	{
-		"fname": "Jon",
-		"lname": "Diom",
-		"startDate": "2/13/94",
-		"street": "79 North Main PO Box 79",
-		"city": "Collinston",
-		"state": "ID",
-		"zip": 93620,
-		"phone": "239-3392",
-		"s_year": 3,
-		"id": "0015"
-	},
-	{
-		"fname": "Jenna",
-		"lname": "Larsen",
-		"startDate": "12/14/93",
-		"street": "PO Box 33 463 West Main Street",
-		"city": "Richards",
-		"state": "ID",
-		"zip": 93637,
-		"phone": "439-2221",
-		"s_year": 3,
-		"id": "0016"
-	},
-	{
-		"fname": "Aonna",
-		"lname": "Saker",
-		"startDate": "3/24/93",
-		"street": "311 East 30 North",
-		"city": "Jonefield",
-		"state": "ID",
-		"zip": 93667,
-		"phone": "369-3336",
-		"s_year": 4,
-		"id": "0017"
-	},
-	{
-		"fname": "Josh",
-		"lname": "Alexander",
-		"startDate": "12/27/93",
-		"street": "322 East 200 South",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "733-0336",
-		"s_year": 3,
-		"id": "0018"
-	},
-	{
-		"fname": "Josie",
-		"lname": "Kane",
-		"startDate": "3/3/93",
-		"street": "64 Canterbury Lane",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "679-0200",
-		"s_year": 4,
-		"id": "0019"
-	},
-	{
-		"fname": "Julius",
-		"lname": "Brooks",
-		"startDate": "6/19/93",
-		"street": "937 Sumac Drive",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "733-6790",
-		"s_year": 4,
-		"id": "0020"
-	},
-	{
-		"fname": "Derrick",
-		"lname": "Dovovan",
-		"startDate": "2/26/93",
-		"street": "373 Edgewood Dr.",
-		"city": "Harrisburg",
-		"state": "ID",
-		"zip": 93632,
-		"phone": "734-3176",
-		"s_year": 4,
-		"id": "0021"
-	},
-	{
-		"fname": "John",
-		"lname": "Hank",
-		"startDate": "1/6/93",
-		"street": "270 West 400 South",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93673,
-		"phone": "734-1643",
-		"s_year": 4,
-		"id": "0022"
-	},
-	{
-		"fname": "Gordon",
-		"lname": "O'Brian",
-		"startDate": "2/3/93",
-		"street": "397 N 300 W",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 94321,
-		"phone": "673-2344",
-		"s_year": 4,
-		"id": "0023"
-	},
-	{
-		"fname": "Thomas",
-		"lname": "Steen",
-		"startDate": "9/1/92",
-		"street": "373 Bringhurst Dr.",
-		"city": "Harrisburg",
-		"state": "ID",
-		"zip": 93632,
-		"phone": "679-4000",
-		"s_year": 4,
-		"id": "0024"
-	},
-	{
-		"fname": "Rick",
-		"lname": "Williams",
-		"startDate": "3/22/93",
-		"street": "114 East Stadium Dr.",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93622,
-		"phone": "797-6914",
-		"s_year": 4,
-		"id": "0025"
-	},
-	{
-		"fname": "Nate",
-		"lname": "Hatch",
-		"startDate": "1/19/94",
-		"street": "179 West 1190 North",
-		"city": "Williamsburg",
-		"state": "ID",
-		"zip": 93679,
-		"phone": "673-4490",
-		"s_year": 3,
-		"id": "0026"
-	}
-]
